@@ -1,28 +1,23 @@
 {
-  "name": "spectacle-ts",
+  "name": "ekg-presentation",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/user-event": "^13.2.1",
-    "@types/jest": "^27.0.1",
     "@types/node": "^16.7.13",
-    "@types/react": "^17.0.20",
-    "@types/react-dom": "^17.0.9",
+    "@types/react": "^18.2.0",
+    "@types/react-dom": "^18.2.0",
     "add": "^2.0.6",
-    "assert": "^2.0.0",
     "beautiful-react-diagrams": "^0.5.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "5.0.0",
-    "spectacle": "^8.5.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "spectacle": "^10.1.4",
     "typescript": "^4.4.2",
     "web-vitals": "^2.1.0",
     "yarn": "^1.22.17"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "./env; react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
