@@ -1,18 +1,20 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import "./index.css";
 import "beautiful-react-diagrams/styles.css";
-import App from "./App";
 import reportWebVitals from "./reportWebVitals";
+import CurrentDeck from "./presentation";
 
-ReactDOM.render(
+function App() {
+  return (<CurrentDeck />);
+}
+
+const root = createRoot(document.getElementById("root")!);
+
+root.render(
   <React.StrictMode>
     <App />
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
