@@ -1,9 +1,10 @@
 import React from "react";
 import { Deck } from "spectacle";
-import { customTheme } from "../../organization.settings";
-import { BaseSlide, Title } from "../../templates";
+import { customTheme } from "../settings";
+import { BaseSlide, Title } from "../templates";
 import deckSettings from "./deck.settings";
 import Diagram, { createSchema, useSchema } from "beautiful-react-diagrams";
+import { template } from "../templates/template";
 
 const DiagramNode = () => (
   <div className="bg-white w-64 p-2 b-1 text-center rounded-xl">
@@ -31,14 +32,14 @@ const diagramSchema = createSchema({
   ],
 });
 
-const MyDeck = () => {
+const Presentation = () => {
   // create diagrams schema
   const [schema, { onChange }] = useSchema(diagramSchema);
   return (
-    <Deck theme={customTheme}>
+    <Deck theme={customTheme} template={template}>
       <Title title={deckSettings.deckTitle} />
       <BaseSlide>
-        <div className="flex justify-between  ">
+        {/* <div className="flex justify-between  ">
           <div>hello 1</div>
           <div>hello 2</div>
           <div>hello 3</div>
@@ -46,9 +47,9 @@ const MyDeck = () => {
         </div>
         <div className="h-96">
           <Diagram schema={schema} onChange={onChange} />
-        </div>
+        </div> */}
       </BaseSlide>
     </Deck>
   );
 };
-export default MyDeck;
+export default Presentation;
