@@ -1,6 +1,5 @@
 import React from "react";
-import { FlexBox, Heading, Slide, Text } from "spectacle";
-import { customTheme, organizationSettings } from "../organization.settings";
+import { Box, FlexBox, FullScreen, Heading, Progress, Slide, Text } from "spectacle";
 
 interface TitleSlideInputs {
   title: string;
@@ -8,24 +7,12 @@ interface TitleSlideInputs {
 }
 
 const TitleSlide: React.FC<TitleSlideInputs> = ({ title, subtitle }) => (
-  <Slide backgroundColor="secondary">
+  <Slide backgroundColor="lightWhite">
     <FlexBox flexDirection="column" justifyContent="center" height="100%">
-      <Heading margin="2px" padding="2px" color="lightWhite">
+      <Heading margin="2px" padding="2px" color="primary">
         {title}
       </Heading>
-      <div
-        style={{
-          marginTop: "20px",
-          height: "4px",
-          width: "120px",
-          backgroundColor: customTheme.colors.lightWhite,
-          borderRadius: "20px",
-        }}
-      />
       <Text textAlign="center">{subtitle}</Text>
-    </FlexBox>
-    <FlexBox alignItems="center">
-      <img src={organizationSettings.logoLightUrl} width="120px" />
     </FlexBox>
   </Slide>
 );
